Add current weather model and update other models

+++ app/src/main/java/com/example/weatherapp/domain/models/TemperatureModel.kt
@@ -1,6 +1,5 @@
 package com.example.weatherapp.domain.models
 
-import com.google.gson.annotations.SerializedName
 
 /**
  * Created by Batuhan Duvarci on 8.12.2021.
@@ -8,8 +7,8 @@ import com.google.gson.annotations.SerializedName
 data class TemperatureModel(
     val day: Double? = null,
     val eve: Double? = null,
-    val max: Double? = null,
-    val min: Double? = null,
+    val max: Int? = null,
+    val min: Int? = null,
     val morn: Double? = null,
     val night: Double? = null
 )
